\name{KG_P1_SampData}
\docType{data}
\alias{KG_P1_SampData}
\title{Simulated sample data for 1000 Genomes Phase 1}
\description{
    An AnnotatedDataFrame with columns sample.id, sex, age, and phenotype,
where the identifiers in sample.id match those in the SeqArray file.
}
\value{An AnnotatedDataFrame}

\usage{KG_P1_SampData}

\keyword{gds}
\keyword{sequencing}
\keyword{genetics}
